minor simplifications
