+++ package.json
@@ -27,7 +27,6 @@
     "preact": "^10.18.1",
     "preact-async-route": "^2.2.1",
     "preact-router": "^4.1.2",
-    "short-uuid": "^4.2.2",
     "simple-peer": "9.11.1",
     "simple-peer-light": "9.10.0",
     "styled-components": "5.3.11",
