+++ src/stateManagers/downloadManagers/realFileManager.ts
@@ -1,5 +1,5 @@
 import { create } from "zustand";
-import { uuidSource } from "../../helpers/helpers";
+import { genId } from "../../helpers/utils";
 
 interface RealFileStore {
   realFiles: { [key: string]: File };
@@ -16,7 +16,7 @@ export const useRealFiles = create<RealFileStore>((set) => ({
 				...(() => {
 					const newFiles: { [key: string]: File } = {};
 					for (const file of files) {
-						newFiles[uuidSource()] = file;
+						newFiles[genId(6)] = file;
 					}
 					return newFiles;
 				})()
